requirements.txt
boto3
dnspython
requests